third_party/blink/web_tests/platform/generic/http/tests/misc/extract-http-content-language-malformed-expected.txt

Test for bug 97929: Extract HTTP Content-Language header.

==> Value extracted from malformed HTTP "Content-Language" header...
PASS window.getComputedStyle(document.getElementById("x")).webkitLocale == '"es"' || window.getComputedStyle(document.getElementById("x")).webkitLocale == "auto" is true
==> Value set by div "lang" tag...
PASS window.getComputedStyle(document.getElementById("y")).webkitLocale is '"ar"'
==> All done...
PASS successfullyParsed is true

TEST COMPLETE
